Switch to refcount logic for kernels >= 5.9
Starting from kernel 5.9+ function 'put_seccomp_filter' has been inlined
and unavailble for hooking. However, internal not-inline function was used
to mitigate the problem. Unfortunately, there is no equivalent counter-part
function for new hook and the old one looks incompatible which we overlooked.
This patch is switching the logic for kernels 5.9+ to custom implementation
of refcount logic and it should address the issue reported as #338